fix: Stopping the process if the version does not exist, and prevent having a wrong fs if the cli is interupted
