+++ src/commands.rs
@@ -1,4 +1,5 @@
 use crate::bun;
+use crate::os;
 use crate::utils;
 use anyhow::Result;
 use lazy_static::lazy_static;
@@ -8,8 +9,8 @@ use std::borrow::Cow;
 #[cfg(not(windows))]
 use std::os::unix::fs::PermissionsExt;
 use std::path::{Path, PathBuf};
-use tokio::fs::{remove_dir_all, File};
-use tokio::{fs, io};
+use tokio::fs;
+use tokio::fs::remove_dir_all;
 use utils::check_folder_exists;
 
 lazy_static! {
@@ -36,14 +37,14 @@ async fn get_bumrc_version() -> Result<String, &'static str> {
 pub async fn use_bun(version: &str) -> Result<()> {
     let version = normalize_version(version);
 
-    let arch = utils::get_architecture();
-
     let active_color: DynColors = "#eea990".parse().unwrap();
     let active_style = owo_colors::Style::new().color(active_color).bold();
 
     let path_with_version = FOLDER_VERSION_BASE.join(&version);
 
     if check_folder_exists(&path_with_version).await {
+        let arch = os::get_architecture();
+
         let bun_used_path = path_with_version.join(format!("bun-{}/bun", arch));
         if let Ok(()) = activate_bun(bun_used_path).await {
             println!(
@@ -59,51 +60,28 @@ pub async fn use_bun(version: &str) -> Result<()> {
 
     println!("Bum - installing bun for version {}...", version);
 
-    let github_bun_download_url = format!(
-        "https://github.com/oven-sh/bun/releases/download/bun-v{}/bun-{}.zip",
-        version, arch
-    );
-
     if fs::metadata(FOLDER_VERSION_BASE.to_owned()).await.is_err() {
         let _ = fs::create_dir_all(FOLDER_VERSION_BASE.to_owned());
     }
 
     let zip_file_path = FOLDER_VERSION_BASE.join(format!("{}.zip", version));
-    println!("{:?}", zip_file_path);
 
-    match utils::download_zip(&github_bun_download_url, &zip_file_path).await {
-        Ok(()) => {
-            let _ = utils::unzip_file(&zip_file_path, &FOLDER_VERSION_BASE).await;
-
-            let bun_used_path = FOLDER_VERSION_BASE
-                .join(&version)
-                .join(format!("bun-{}/bun", arch));
-
-            match activate_bun(bun_used_path).await {
-                Ok(()) => println!(
-                    "Bun {} is activated.",
-                    format!("v{}", version).style(active_style)
-                ),
-                _ => println!("Failed"),
-            }
-        }
-        Err(err) => eprintln!("Error: {}", err),
+    match bun::download_version_to(&version, &zip_file_path).await {
+        Ok(bun_path) => match activate_bun(bun_path).await {
+            Ok(()) => println!(
+                "Bun {} is activated.",
+                format!("v{}", version).style(active_style)
+            ),
+            e => eprintln!("Failed {e:?}"),
+        },
+        Err(err) => eprintln!("{err}"),
     }
 
     Ok(())
 }
 
 pub async fn activate_bun(bun_used_path: PathBuf) -> Result<()> {
-    fs::remove_file(BUN_BIN_PATH.to_owned())
-        .await
-        .expect("Failed to delete the current bun bin");
-
-    let mut file_to_copy = File::open(bun_used_path).await.unwrap();
-    let mut file_target = File::create(BUN_BIN_PATH.to_owned()).await.unwrap();
-
-    io::copy(&mut file_to_copy, &mut file_target)
-        .await
-        .expect("Faield to copy the bun bin");
+    fs::copy(bun_used_path, BUN_BIN_PATH.to_owned()).await?;
 
     let metadata = fs::metadata(BUN_BIN_PATH.to_owned()).await?;
